fix: raising import error when not install using discord extras
